requirements.txt
# AutoRegressive model requirements
numpy>=1.20.0
scikit-learn>=1.0.0
statsmodels>=0.13.0
pyyaml>=6.0
matplotlib>=3.5.0
ray>=2.0.0
